Fix compiler error and add skin number change
The compiler had a chance to give an error when running the "Update Files" option if a character didn't have any NPCs set up. This was fixed.

Also, shared NPC files, if they're skins/huds/3D heads, can now use a different number for the NPC. They'll be properly renamed and hex edited if they're different. The existing compileInfo.xml files have been updated to support this, and the compiler has been updated as well.

"Update Files" option ran as a test.